Modularize ovsp4rt.cc
- Updated ovs submodule reference.

- Moved utility functions to ovsp4rt_utils.

- Separated target-independent functions from functions that are
  specific to DPDK or ES2K.

- Separated public functions (ones exposed to openvswitch through
  ovs-p4rt.h) from private functions (ones internal to ovs-p4rt).

- Added #else #error clauses to compound conditionals

- Deleted gratuitous return statements
